+++ src/shillelagh/backends/apsw/dialects/base.py
@@ -1,5 +1,8 @@
+"""
+A SQLALchemy dialect.
+"""
+
 # pylint: disable=protected-access, abstract-method
-"""A SQLALchemy dialect."""
 
 from typing import Any, Dict, List, Optional, Tuple, cast
 
@@ -102,7 +105,6 @@ class APSWDialect(SQLiteDialect):
         connection: _ConnectionFairy,
         table_name: str,
         schema: Optional[str] = None,
-        info_cache: Optional[Dict[Any, Any]] = None,
         **kwargs: Any,
     ) -> bool:
         """
@@ -111,7 +113,14 @@ class APSWDialect(SQLiteDialect):
         try:
             get_adapter_for_table_name(connection, table_name)
         except ProgrammingError:
-            return False
+            return bool(
+                super().has_table(
+                    connection,
+                    table_name,
+                    schema,
+                    **kwargs,  # pylint: disable=unused-argument
+                ),
+            )
         return True
 
     # needed for SQLAlchemy